+++ test_muacrypt/test_cmdline.py
@@ -7,6 +7,7 @@ import re
 import six
 import pytest
 from muacrypt import mime
+from .test_account import gen_ac_mail_msg
 
 
 def test_help(cmd):
@@ -91,11 +92,22 @@ class TestProcessIncoming:
 
     def test_process_incoming_no_autocrypt(self, mycmd, datadir):
         mycmd.run_ok(["add-account", "--email-regex=b@b.org"])
+        mycmd.run_ok(["peerstate", "a@a.org"])
         msg = mime.gen_mail_msg(From="Alice <a@a.org>", To=["b@b.org"], _dto=True)
         mycmd.run_ok(["process-incoming"], """
             *processed*default*no*Autocrypt*header*
         """, input=msg.as_string())
-        mycmd.run_ok(["status"])
+        mycmd.run_ok(["peerstate", "a@a.org"])
+
+    def test_peer_with_ac_keys(self, mycmd, datadir):
+        mycmd.run_ok(["add-account", "-a", "acc1", "--email-regex=a@a.org"])
+        mycmd.run_ok(["add-account", "-a", "acc2", "--email-regex=b@b.org"])
+        acc1, acc2 = mycmd.get_account("acc1"), mycmd.get_account("acc2")
+        acc1.addr, acc2.addr = "a@a.org", "b@b.org"
+        acc2.process_incoming(gen_ac_mail_msg(acc1, acc2))
+        mycmd.run_ok(["peerstate", "-a", "acc2", "a@a.org"])
+        acc1.process_incoming(gen_ac_mail_msg(acc2, acc1))
+        mycmd.run_ok(["peerstate", "-a", "acc1", "b@b.org"])
 
 
 class TestAccountCommands: